I started with a main function.
So here's the program to check whether the string entered by the user is a palindrome or not. I made a main function just to get things started. The user inputs the length of the string first. Then the string is entered. There's an if-else which checks whether the string is a palindrome by calling a isPalindrome(), which I will be making in a while.